added node block kernel 2, fixed SAM example, fixed USB clock issue, fixed SDL variant

+++ desktop/cores/sdl/KLST_SDL-adapter.cpp
@@ -28,7 +28,7 @@ KLST_Simulator mSimulator;
 
 void klangstrom::begin_serial_debug(bool pWaitForSerial, uint32_t pBaudRate) {}
 
-void klangstrom::option(uint8_t pOption, uint32_t pValue) {}
+void klangstrom::option(uint8_t pOption, float pValue) {}
 
 void klangstrom::beats_per_minute(float pBPM) { klangstrom_arduino_beats_per_minute(pBPM); }
